requirements.txt
langchain
langgraph
langchain-openai
langchain-google-genai
feedparser
requests
pandas
pydantic
python-dotenv
